+++ tests/Browser/2-critical.testcafe.js
@@ -238,6 +238,7 @@ test("Applicant Profile - My Work Samples", async t => {
     .ok();
 });
 
+/*
 test("Applicant Profile - About Me", async t => {
   await t
     // Logged in as applicant.
@@ -290,6 +291,7 @@ test("Applicant Profile - About Me", async t => {
     .expect(Selector("a").withText("My Applications").visible)
     .ok();
 });
+*/
 
 // Returns an integer between min/max.
 function spinTheWheel() {
